+++ lib/app_with_flutter_bloc/features/title/bloc/title_state.dart
@@ -1,0 +1,23 @@
+part of 'title_bloc.dart';
+
+abstract class TitleState extends Equatable {
+  const TitleState();
+}
+
+final class TitleInitial extends TitleState {
+  @override
+  List<Object> get props => [];
+}
+
+final class TitleWithValue extends TitleState {
+  const TitleWithValue({required this.title});
+  final String title;
+
+  @override
+  String toString() {
+    return 'TitleWithValue - $title';
+  }
+
+  @override
+  List<Object> get props => [title];
+}