+++ lib/app_with_flutter_bloc/presentions/widgets/o_b_s_action_buttons_mobile.dart
@@ -1,10 +1,7 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:obs_for_sama/app_with_flutter_bloc/features/server/bloc/server_bloc.dart';
 import 'package:obs_for_sama/app_with_flutter_bloc/presentions/widgets/go_to_setting_page_button.dart';
-import 'package:obs_for_sama/app_with_flutter_bloc/presentions/widgets/o_b_s_reload_button.dart';
 import 'package:obs_for_sama/app_with_flutter_bloc/presentions/widgets/o_b_s_toogle_sound_button.dart';
 import 'package:obs_for_sama/app_with_flutter_bloc/presentions/widgets/o_b_s_toogle_stream_button.dart';
 import 'package:obs_for_sama/core/index.dart';
@@ -39,25 +36,26 @@ class OBSActionButtonsMobile extends StatelessWidget {
                           children: [
                             Row(
                               children: [
-                                /// CONNECT OBS
                                 Expanded(
                                   child: Padding(
                                     padding: const EdgeInsets.all(8),
-                                    child: Platform.isAndroid
-                                        ? const GoToSettingPageButton(
-                                            key: ValueKey<String>('Android Button connection'),
-                                          )
-                                        : const GoToSettingPageButton(
-                                            key: ValueKey<String>('IOS Button connection'),
-                                          ),
-                                  ),
-                                ),
-
-                                /// REFRESH
-                                const Expanded(
-                                  child: Padding(
-                                    padding: EdgeInsets.all(8),
-                                    child: OBSReloadButton(),
+                                    child: SizedBox(
+                                      width: 30,
+                                      child: RSIButtonOutlined(
+                                        onTap: Scaffold.of(context).openDrawer,
+                                        edgeClipper: const RSIEdgeClipper(
+                                          edgeRightTop: true,
+                                          // edgeLeftBottom: true,
+                                          // edgeRightBottom: true,
+                                          edgeLeftTop: true,
+                                        ),
+                                        child: Icon(
+                                          Icons.menu,
+                                          size: 40,
+                                          color: Theme.of(context).colorScheme.outline,
+                                        ),
+                                      ),
+                                    ),
                                   ),
                                 ),
 
@@ -65,7 +63,10 @@ class OBSActionButtonsMobile extends StatelessWidget {
                                 const Expanded(
                                   child: Padding(
                                     padding: EdgeInsets.all(8),
-                                    child: OBSToogleSoundButton(),
+                                    child: SizedBox(
+                                      width: 30,
+                                      child: OBSToogleSoundButton(),
+                                    ),
                                   ),
                                 ),
                               ],